+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 ## --- [4.0.15] - 2022/10/02
 ### New features
-TBD
+- Base Theme Switching (Dark, Light, Default) 🤩🎨 ([Merge Request](https://gitlab.com/crafty-controller/crafty-4/-/merge_requests/471))
 ### Bug fixes
 - Fix traceback on basic schedule with "days" interval ([Merge Request](https://gitlab.com/crafty-controller/crafty-4/-/merge_requests/469))
 - Fix bad method call with API stdin ([Merge Request](https://gitlab.com/crafty-controller/crafty-4/-/merge_requests/470))<br>